requirements.txt
torch==2.4.1
torchvision==0.19.1
ultralytics==8.3.0
streamlit==1.39.0
pillow==10.4.0
numpy==1.26.4
pandas==2.2.2
tqdm==4.66.5
opencv-python==4.10.0.84
